2.Titanic - Machine Learning from Disaster
In this file, First I did is cleaned the data and then used a visualization tool to fetch meaningful information. After that, I did standardization of data so I can build a model to predict who survive in Titanic Disaster. And I build a basic model to predict survival in Titanic Disaster.